requirements.txt
asgiref==3.6.0
Django==4.1.7
Pillow==9.4.0
python-environ==0.4.54
sqlparse==0.4.3
tzdata==2022.7
whitenoise==6.6.0
django-markdown-deux